[use-package] Fix usage of :after
:after defers loading one package until its dependencies have been
loaded, (push semantics) but I was using it in cases where I wanted
pull semantics (if you are loading this package, also ensure these
other packages are loaded).

Fixed by utilising :autoload on the depended upon package to expose an
interface that will trigger the package to be loaded, rather than
defer loading the package that depends on it.