[evil/org] Bindings to insert items before and after
Add new elements to the outline (headings or checkboxes), before or
after the element at point.  The element inserted depends on the
element at point (i.e. it matches: task or no task, heading or list)
and it respects structure (i.e. if a heading or bulleted list has
children, the entry goes after them).

Also introduces an evil binding for archive, which conflicts with the
old binding for refile which has since been moved.